Split function method template resolver tests
